backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
Django==3.2
django-filter==23.2
django-colorfield==0.9.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
djoser==2.2
drf-extra-fields==3.4.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==1.7.0
install==1.3.5
isort==5.11.5
mccabe==0.7.0
pep8-naming==0.13.3
PyYAML==6.0
Pillow==10.0.0
psycopg2-binary==2.9.6
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
zipp==3.15.0
webcolors==1.11.1